Drawdowns for Optiwells configuration and pump discharges

Calculate drawdowns for given Optiwells configuration and pump discharges

owGetDrawdowns(owConfiguration, Q, times = NULL, solution = NULL,

    to.matrix = TRUE, ...)

Arguments

owConfiguration
Optiwells configuration as e.g. retrieved by owRandomConfiguration.

Q
vector of discharges at the wells in the order of wells in

owConfiguration$wellfield

times
vector of times for which drawdowns are to be calculated

solution
List of solution parameters as retrieved by

wtConfigureSolution. If not specified, a default

configuration, as retrieved by wtDefaultConfigurationSolution

is used.

to.matrix
if TRUE, the results returned by wtRunConfiguration that

come in "long" form (the drawdowns for each well appear in blocks one

below each other with a column "WELL" indicating the well name) are

converted to "matrix" (= wide) form in which the calculated drawdowns

appear in columns one beside the other.

…
additional arguments passed to wtRunConfiguration, such as

show.results or dbg, see there.

Value

list with as many elements as there are wells defined in the Optiwells

configuration (list element wellfield of owConfiguration). For each

pumping well (well with Q > 0) the list element at the corresponding

position contains a data frame holding the drawdowns calculated by WTAQ

with that well pumping alone and the other wells being observation wells.

For each non-pumping well (well with Q = 0) the list element will be NULL.

This list can then be used to calculate the superposition of drawdowns.

See also