Generation of a random wellfield

Generation of a random wellfield

owRandomWellfield(numberOfWells, rmean, z1mean, z2mean, swmean, 
    rsd = 0.1 * rmean, z1sd = 0.1 * z1mean, z2sd = 0.1 * z2mean, 
    swsd = 0.1 * swmean, digits = 1)

Arguments

numberOfWells

number of wells in the well field (must be a number between 1 and 25).

rmean

mean value of normal distribution used to generate random values for 'r' (well radius)

z1mean

mean value of normal distribution used to generate random values for 'z1' (depth of begin of screen)

z2mean

mean value of normal distribution used to generate random values for 'z2' (depth of end of screen)

swmean

mean value of normal distribution used to generate random values for 'sw' (well-bore skin parameter)

rsd

standard deviation of normal distribution used to generate random values for 'r' (well radius). Default: 0.1 * rmean

z1sd

standard deviation of normal distribution used to generate random values for 'z1'-values' (depth of begin of screen). Default: 0.1 * z1mean

z2sd

standard deviation of normal distribution used to generate random values for 'z2'-values' (depth of end of screen). Default: 0.1 * z2mean

swsd

standard deviation of normal distribution used to generate random values for 'sw' (well-bore skin parameter). Default: 0.1 * swmean

digits

number of decimal digits for values of r, z1, z2. Default: 1

Value

data frame with each row representing a well of the well field and the columns representing the well properties: wellName (well name), x (x-coordinate of well), y-coordinate of well, r

(well radius), z1 (depth of begin of well screen), z2 (depth of end of well screen), sw (well-bore skin parameter)

Author

Hauke Sonnenberg