Define parameter values for WTAQ model run

Define parameter values for WTAQ model run

wtConfigure(general = wtConfigureGeneral(), aquifer = wtConfigureAquifer(), 
    drainage = wtConfigureDrainage(), times = wtConfigureTimes(), 
    solution = wtConfigureSolution(), pumpwell = wtConfigurePumpwell(), 
    obswells = list(wtConfigureObservationWell(obname = "obs1"), 
        wtConfigureObservationWell(obname = "obs2", r = 50)))

Arguments

general

List of general parameters as retrieved by wtConfigureGeneral.

aquifer

List of aquifer parameters as retrieved by wtConfigureAquifer.

drainage

List of drainage parameters as retrieved by wtConfigureDrainage.

times

List of simulation time parameters as retrieved by wtConfigureTimes.

solution

List of solution parameters as retrieved by wtConfigureSolution.

pumpwell

List of pumped well parameters as retrieved by wtConfigurePumpwell.

obswells

List of lists of observation well parameters each of which can be retrieved by wtConfigureObservationWell.

Value

list with elements general, aquifer, drainage, times, solution, pumpwell, obswells, representing a full WTAQ configuration. For the different elements see the descriptions in the Arguments section.

Author

Hauke Sonnenberg

See also