Aquifer parameter values

Define aquifer-related WTAQ parameters

wtConfigureAquifer(aqtype = "WATER TABLE", bb = 20, hkr = 0.0041,
    hkz = 0.00174, ss = 3.76e-05, sy = 0.25)

Arguments

aqtype
Type of aquifer being simulated. Two options are provided: AQTYPE = CONFINED or AQTYPE = WATER TABLE

bb
Thickness or saturated thickness of aquifer at beginning of simulation, in units of length.

hkr
Horizontal hydraulic conductivity of aquifer, in units of length per time.

hkz
Vertical hydraulic conductivity of aquifer, in units of length per time.

ss
Specific storage of aquifer, in units of inverse length.

sy
Specific yield of aquifer, dimensionless. Enter 0 if AQTYPE = CONFINED.

Value

list with elements aqtype, bb, hkr, hkz, ss, sy, representing WTAQ parameters related to the aquifer. See descriptions in Arguments section.

See also