Observation well parameter values

Define WTAQ parameters related to an obervation well

wtConfigureObservationWell(obname = "obs1", r = 30, iows = 0,
    idpr = 0, rp = ifelse(idpr == 0, 0, 0.1), z1 = ifelse(iows ==
        2, 0, 5), z2 = ifelse(iows == 2, 0, 10), xll = ifelse(idpr ==
        0, 0, z2 - z1), zp = ifelse(iows == 2, 3, 0), tsobs = data.frame(t = 60 *
        1:3, dd = c(0.4, 0.6, 0.7)), irun = 1)

Arguments

obname
Name of observation well or piezometer; up to 10 characters in length.

r
Radial distance from axis of pumped well to observation well or piezometer, in units of length.

iows
Type of observation well or piezometer: IOWS = 0: Partially penetrating observation well. IOWS = 1: Fully penetrating observation well. IOWS = 2: Observation piezometer. Default: 0

idpr
Options for delayed response of observation well. IDPR = 0: No delayed response. IDPR = 1: Delayed response. Default: 0

rp
Inside radius of the observation well (or piezometer) standpipe in the interval over which water levels are changing during pumping, in units of length. Enter 0 if IDPR = 0. Default: 0.1, but 0 if idpr = 0.

z1
Depth below top of aquifer or initial water table to the top of screened interval of observation well, in units of length. Use for IOWS = 0 or 1. Enter 0 if IOWS = 2. Default: 5.0, but 0 if iows = 2.

z2
Depth below top of aquifer or initial water table to the bottom of screened interval of observation well, in units of length. Use for IOWS = 0 or 1. Enter 0 if IOWS = 2. Default: 10.0, but 0 if iows = 2.

xll
Length of screened interval of observation well or piezometer, in units of length. Must be 0 if IDPR = 0. Default: z2 - z1, but 0 if idpr = 0.

zp
Depth below top of aquifer or initial water table to center of piezometer, in units of length. Use for IOWS = 2. Enter 0.0 if IOWS = 0 or 1. Default: 3.0, but 0.0 if iows not equal to 2.

tsobs
data.frame with column t containing the user-specified times for which drawdown at the observation well or piezometer will be calculated. If the data.frame has no rows, no drawdowns are calculated for the observation well or piezometer. The data.frame can (optionally) contain the measured drawdown for the observation well or piezometer at the corresponding times in an additional column dd.

irun
Option to suppress drawdown calculations for the observation well or piezometer. Allows user to specify time-drawdown data, but those data are ignored during the simulation. Options are: IRUN = 0: Drawdowns not calculated. IRUN = 1: Drawdowns calculated.

Value

list with elements irun, obname, iows, idpr, r, z1, z2, zp, rp, xll, tsobs, representing WTAQ parameters related to an observation well. See description in Arguments section.

See also